+++ accessories/humiditySensor.js
@@ -55,6 +55,21 @@ class HumiditySensorAccessory extends HumidifierAccessory {
   //Method inhertied but not required
   async updateDeviceState () { return;}
   
+  getBatteryAlert (callback) {
+    const { config, name, state, log, debug } = this;
+
+    const batteryAlert = state.batteryLevel <= 20? Characteristic.StatusLowBattery.BATTERY_LEVEL_LOW : Characteristic.StatusLowBattery.BATTERY_LEVEL_NORMAL;
+    if (debug) log(`\x1b[34m[DEBUG]\x1b[0m ${name} Battery Level:`,state.batteryLevel,'Alert:',batteryAlert);
+
+    callback(null, batteryAlert);
+  }
+  
+  getBatteryLevel (callback) {
+    const { config, name, state, log, debug } = this;
+    if (debug) log(`\x1b[34m[DEBUG]\x1b[0m ${name} Battery Level:`,state.batteryLevel);
+    callback(null, parseFloat(state.batteryLevel));
+  }
+
   // Service Manager Setup
   setupServiceManager () {
     const { config, name, serviceManagerType } = this;
@@ -67,6 +82,22 @@ class HumiditySensorAccessory extends HumidifierAccessory {
       method: this.getCurrentHumidity,
       bind: this
     });
+
+    if (config.batteryAlerts){
+      this.serviceManager.addGetCharacteristic({
+        name: 'batteryAlert',
+        type: Characteristic.StatusLowBattery,
+        method: this.getBatteryAlert,
+        bind: this
+      })
+      
+      this.serviceManager.addGetCharacteristic({
+        name: 'batteryLevel',
+        type: Characteristic.BatteryLevel,
+        method: this.getBatteryLevel,
+        bind: this
+      })
+    };
   }
 }