+++ accessories/temperatureSensor.js
@@ -15,7 +15,6 @@ class TemperatureSensorAccessory extends AirconAccessory {
     super(log, config, serviceManagerType);
 
     this.temperatureCallbackQueue = {};
-    this.monitorTemperature();
   }
 
   setDefaults () {
@@ -42,6 +41,21 @@ class TemperatureSensorAccessory extends AirconAccessory {
     super.reset();
   }
 
+  getBatteryAlert (callback) {
+    const { config, name, state, log, debug } = this;
+
+    const batteryAlert = state.batteryLevel <= 20? Characteristic.StatusLowBattery.BATTERY_LEVEL_LOW : Characteristic.StatusLowBattery.BATTERY_LEVEL_NORMAL;
+    if (debug) log(`\x1b[34m[DEBUG]\x1b[0m ${name} Battery Alert:`,batteryAlert);
+
+    callback(null, batteryAlert);
+  }
+  
+  getBatteryLevel (callback) {
+    const { config, name, state, log, debug } = this;
+    if (debug) log(`\x1b[34m[DEBUG]\x1b[0m ${name} Battery Level:`,state.batteryLevel);
+    callback(null, parseFloat(state.batteryLevel));
+  }
+  
   // Service Manager Setup
 
   setupServiceManager () {
@@ -64,7 +78,23 @@ class TemperatureSensorAccessory extends AirconAccessory {
         bind: this
       })
     };
-
+    
+    if (config.batteryAlerts){
+      this.serviceManager.addGetCharacteristic({
+        name: 'batteryAlert',
+        type: Characteristic.StatusLowBattery,
+        method: this.getBatteryAlert,
+        bind: this
+      })
+      
+      this.serviceManager.addGetCharacteristic({
+        name: 'batteryLevel',
+        type: Characteristic.BatteryLevel,
+        method: this.getBatteryLevel,
+        bind: this
+      })
+    };
+    
     this.serviceManager.addGetCharacteristic({
       name: 'temperatureDisplayUnits',
       type: Characteristic.TemperatureDisplayUnits,