+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changes
 
+## 4.4.0
+ - [Improvement] Updated all dependencies to remove some security vulnerabilities 
+ - [Improvement] Added Low Battery Alerts to Humidity and Temperature sensors (using battery:XX in readings from files)
+ - [Improvement] Extended Temperature and Humidity readings from files to support temperature:XX, humidity:XX, and battery:XX values on each line
+ - [Improvement] Removed file and w1 Temperature readings being forced to 10 minutes. Now 1 minute minimum and adjustable
+ - [Improvement] Added MQTT support to AC and Temperature sensor using temperature,humidity, and battery identifiers
+ - [Improvement] Added HeaterCooler accessory option. Refer to [this document](https://github.com/kiwi-cam/homebridge-broadlink-rm/blob/kiwi-cam-beta-1/docs/heater-cooler.md) to read @newt10's work here.
+ - [Improvement] Integrated the platform helper module to improve maintainability.
+ - [Improvement] MQTT will update when published so frequent refreshes aren't required ( 10 minute default as a fallback )
+ - [Bug] Fixed duplicate monitorTemperature calls from Temperature Sensor accessories
+ - [Bug] Fixed SIGTERM when unexpected packet received
+
 ## 4.3.8
 - [Improvement] Added Keep Alive packet to RM Devices to avoid reboots when the cloud is unreachable.
 - [Improvement] **Added RF Learning** steps from #45